Enhances certificate installation and UI
Improves the certificate installation process by adding the certificate to both the Trusted Root and Intermediate stores.

Adds a button to the UI to open the output directory for easier access to generated files.

Updates default values in the UI.

Adds VS Code task configurations for build and run.

Adds the output directory to the .gitignore file.